Use integrate orbit allocator as zmalloc
A try of replacing global allocator with orbit allocator.

It turns out to not provide good performance (qps 36% drop, lat. 60%
increase).  We should provide the functionality to clear CoW flags after
the run (as the contrary flag to incremental-snapshotting).

For now we will not use this version, and fallback to replace limited
number of allocation sites for better performance.